Fix test suite for older Perls
The set of names supported by charnames changed between versions,
and we ass-u-me-d that the latest set was supported, instead of
the set from the oldest Perl we support (5.10.0). This fixes the
test to only use names from the set in 5.10. Thanks to Getty for
notifying me of the error.