Use Import::Into instead of a home-grown "solution"
I had to work around mis-parse with Import::Into and open pragma
where "open->import::into(...)" was being parsed as a call to the
function "open" instead of a method call on the open package. But
after reading Import::Into's documentation (and code) further, I
realized that speciyfing open as a string should solve the issue.

RT #78299: open->import::into parsed as a call to the function open

Fixes GH #10: Use Import::Into