Avoid creating files called SCALAR(0x...) in the test suite
Prior to perl v5.14.0, doing `open my $out, '>:raw', \my $mem`
would create a file with the name of the stringification of
the scalar reference in arg 3 - SCALAR(0x...). Without the IO
mode, an in-memory buffer would be opened as expected. By
omitting the IO mode, and applying the ':raw' binmode after
the open succeeded, we avoid creating such files when perl
exhibits this bug.