Mimic readdir's context-sensitivity
CORE::readdir is context-sensitive: in list context, it returns
all the files, which we emulated with utf8 goodness. But in
scalar context, it returns the next file, rather than the number
of files. Mimic this behaviour in utf8::all's wrapper function.

Thanks for Leon Timmermans for catching this omission.

Fixes GH #11