Remove autodie Conflicts; bail out if autodie <= 2.11
utf8::all is broken when used with autodie <= 2.11. This was
previously specified with Conflicts, but that seems to have the
wrong semantics. Since autodie 2.12 was released over a year ago
(June 26, 2012), this should no longer be a problem, so I am
simply removing the Conflicts relationship.

The conflict is still noted in documentation. We also manually
test the autodie version at install time (in t/autodie.t) and
bail out, halting all further testing, if it isn't satisfactory.

Fixes gh-22