Warn instead of bailing out when autodie is too old
It turns out that even today LOTS AND LOTS of people and
smoke testers have old autodies (which actually makes sense)
so bailing out is actually not the right thing to do. It
causes tons of red in cpantesters, and is just super
unhelpful. Humans should be warned to upgrade, and this
does so.

Fixes gh-26 and gh-22